requirements.txt
pandas>=2.2.0
rapidfuzz>=3.9.0
phonenumbers>=8.13.45
email_validator>=2.2.0
pydantic>=2.7.1
PyYAML>=6.0.1
python-dotenv>=1.0.1
great_expectations>=0.18.12
prefect>=3.0.0
httpx>=0.27.2
polars>=1.6.0
duckdb>=1.0.0
